Создает кастомные фильтры. Подключает в виды компоненты Url, Html. Собирает связанные классы в одну директорию и пространство имен.
