+++ frontend/controllers/TasksController.php
@@ -5,8 +5,8 @@ declare(strict_types = 1);
 namespace frontend\controllers;
 
 use yii\web\Controller;
-use frontend\models\TaskService;
-use frontend\models\TaskSearchForm;
+use frontend\models\task\TaskService;
+use frontend\models\task\TaskSearchForm;
 
 class TasksController extends Controller
 {
@@ -20,9 +20,8 @@ class TasksController extends Controller
 
     public function actionIndex()
     {
-        $tasks = $this->service->getTasks(
-            $searchForm = new TaskSearchForm() //а мне первый наставник говорил, что присвоения в условной конструкции - плохая практика. Понимаю, что это не условная конструкция. Но очень похоже не такой случай, хм..
-        );
+        $searchForm = new TaskSearchForm();
+        $tasks = $this->service->getTasks($searchForm);
 
         return $this->render('index', compact('tasks', 'searchForm'));
     }
