+++ frontend/models/cities/Cities.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace frontend\models;
+namespace frontend\models\cities;
 
 use Yii;
