+++ frontend/models/responses/Responses.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace frontend\models;
+namespace frontend\models\responses;
 
 use Yii;
