+++ frontend/models/reviews/Reviews.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace frontend\models;
+namespace frontend\models\reviews;
 
 use Yii;
