+++ frontend/models/specializations/Specializations.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace frontend\models;
+namespace frontend\models\specializations;
 
 use Yii;
