+++ frontend/models/task/TaskHelpfulFiles.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace frontend\models;
+namespace frontend\models\task;
 
 use Yii;
